pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool]

[tool.black]
line-length = 120
target-version = ["py39"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "pass",
    "if 0:",
    "if __name__ == .__main__.:",
    "nocov",
    "if TYPE_CHECKING:",
]
fail_under = 80
show_missing = true

[tool.coverage.run]
branch = true
omit = [
    "tests/*"
]

[tool.isort]
combine_as_imports = "true"
force_grid_wrap = 0
include_trailing_comma = "true"
known_first_party = "src"
line_length = 120
multi_line_output = 3

[tool.mypy]
disallow_untyped_defs = true
follow_imports = "silent"
ignore_missing_imports = true
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.poetry]
authors = ["Szymon Miks <miks.szymon@gmail.com>"]
description = ""
name = "hexagonal-architecture-python"
packages = [{include = "hexagonal_architecture_python"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
boto3 = "^1.25.2"
dropbox = "^11.35.0"
email-validator = "^1.3.0"
fastapi = "^0.85.1"
kink = "^0.6.5"
mypy-boto3-s3 = "^1.24.76"
mypy-boto3-ses = "^1.24.36.post1"
mypy-boto3-sns = "^1.24.68"
pymongo = "^4.2.0"
python = "^3.10"
python-dateutil = "^2.8.2"
requests = "^2.28.1"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
ipython = "^8.5.0"
isort = "^5.10.1"
moto = "^4.0.8"
mypy = "^0.982"
pre-commit = "^2.20.0"
pylint = "^2.15.4"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
toml-sort = "^0.20.1"

[tool.pylint.BASIC]
good-names = "id,i,j,k"

[tool.pylint.DESIGN]
max-args = 5
max-attributes = 8
min-public-methods = 1

[tool.pylint.FORMAT]
max-line-length = 120

[tool.pylint."MESSAGES CONTROL"]
disable = "missing-docstring, line-too-long, logging-fstring-interpolation, duplicate-code"

[tool.pylint.MISCELLANEOUS]
notes = "XXX"

[tool.pylint.SIMILARITIES]
ignore-comments = "yes"
ignore-docstrings = "yes"
ignore-imports = "yes"
min-similarity-lines = 6

[tool.pytest.ini_options]
addopts = "-v --cov=src --cov-report term-missing --no-cov-on-fail"
testpaths = ["tests"]
